+++ cmd/internal/storage/bigquery/client/client.go
