+++ cmd/internal/storage/bigquery/client/example/example.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/SeanDolphin/bqschema"
-	"github.com/google/cadvisor/storage/bigquery/client"
+	"github.com/google/cadvisor/cmd/internal/storage/bigquery/client"
 )
 
 type container struct {
