+++ cmd/internal/storage/influxdb/influxdb_test.go
@@ -25,9 +25,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/google/cadvisor/cmd/internal/storage/test"
 	info "github.com/google/cadvisor/info/v1"
 	"github.com/google/cadvisor/storage"
-	"github.com/google/cadvisor/storage/test"
 
 	influxdb "github.com/influxdb/influxdb/client"
 	"github.com/stretchr/testify/assert"
