+++ cmd/internal/storage/stdout/stdout.go
